fix mint liquidity validators and slplit tests to speed up test run
